add launcher 6

+++ launcher/start.sh
@@ -10,22 +10,27 @@ if [[ -f "homio.properties" ]]; then
     done < "homio.properties"
 fi
 
+mkdir -p $root_path
 echo "root_path: '$root_path'"
 
 java_path=$(command -v java)
 if [[ -z "$java_path" || "$(java -version 2>&1 | awk 'NR==1{print $3}' | cut -d'"' -f2)" != "17" ]]; then
-    java_path="$root_path/jdk-17.0.0/bin/java"
-    echo "Java not installed or version is not 17. Installing..."
+	java_path="$root_path/jdk-17.0.7+7-jre/bin/java"
+    if [ -x "$java_path" ]; then
+	   echo "Java is installed at path $java_path"
+	else
+      echo "Java not installed or version is not 17. Installing..."
 
-    # Download Java 17
-    wget -O "$root_path/jre.tar.gz" 'https://github.com/adoptium/temurin17-binaries/releases/download/jdk-17.0.7+7/OpenJDK17U-jre_x64_linux_hotspot_17.0.7_7.tar.gz'
+      # Download Java 17
+      wget -O "$root_path/jre.tar.gz" 'https://github.com/adoptium/temurin17-binaries/releases/download/jdk-17.0.7+7/OpenJDK17U-jre_x64_linux_hotspot_17.0.7_7.tar.gz'
 
-    # Extract Java archive
-    tar -xf "$root_path/jre.tar.gz" -C "$root_path"
+     # Extract Java archive
+      tar xzf "$root_path/jre.tar.gz" -C "$root_path"
 
-    # Remove the Java archive
-    rm "$root_path/jre.tar.gz"
-    echo "Java 17 has been installed."
+      # Remove the Java archive
+      rm "$root_path/jre.tar.gz"
+      echo "Java 17 has been installed  to $java_path"
+	fi
 else
     echo "Java 17 already installed."
 fi
@@ -49,6 +54,7 @@ else
     app="homio-launcher.jar"
 fi
 
+echo "Run $java_path -jar $root_path/$app"
 "$java_path" -jar "$root_path/$app"
 exit_code=$?
